+++ src/core/framework/network/client/SocketClientHelper.h
@@ -22,7 +22,7 @@ public:
     
     virtual ~SocketClientHelper();
     
-    virtual void processSpecialPacket(uint32_t packetType, InputMemoryBitStream& inInputStream, MachineAddress* inFromAddress);
+    virtual void processSpecialPacket(uint8_t packetType, InputMemoryBitStream& inInputStream, MachineAddress* inFromAddress);
     
     virtual void handleUninitialized();
