+++ src/core/framework/network/steam/NGSteamServerHelper.h
@@ -27,7 +27,7 @@ public:
     
     virtual void processIncomingPackets();
     
-    virtual void processSpecialPacket(uint32_t packetType, InputMemoryBitStream& inInputStream, MachineAddress* inFromAddress);
+    virtual void processSpecialPacket(uint8_t packetType, InputMemoryBitStream& inInputStream, MachineAddress* inFromAddress);
     
     virtual void onClientDisconnected(ClientProxy* clientProxy);
